Got rid of the bin system; the new system is simpler and more robust.
The bins turned out to suffer fundamentally from the fact that the
boundaries can be at arbitrary times that can bunch up two snapshots that
really ought to both be kept.

The next algorithm keeps all the snapshots from the last 24 hours, then
the first snapshot of each day for the last 31 days, then the first
snapshot of each month forever.